管理：ユーザー一覧：編集画面　新規登録処理を追加、画面の横スクロールを削除、郵便番号のバリデーションを修正

+++ app/Http/Requests/Manage/User/UpdateRequest.php
@@ -18,7 +18,7 @@ class UpdateRequest extends BaseRequest
             'userCompany.chargeMail' => ['nullable','email'],
             'userCompany.companyName' => ['nullable','max:40'],
             'userCompany.companyId' => ['required','regex:/^[!-~]+$/','max:5'],
-            'userCompany.postCode' => ['nullable', 'digits:7', 'integer'],
+            'userCompany.postCode' => ['nullable', 'digits:7', 'numeric'],
             'userCompany.address' => ['nullable','max:50'],
             'userCompany.tel' => ['nullable','regex:/^[0-9-]+$/','max:20'],
             'userCompany.staffName' => ['nullable','max:20'],
